+++ NEWS.md
@@ -14,6 +14,7 @@ for more information)
 * Fixed a bug where CPOE was NA when targeting players with apostrophe in last name
 * Added the function `decode_player_ids()` to decode all player IDs to the 
 commonly known GSIS ID format (00-00xxxxx)
+* `fast_scraper_roster()` is finally back!
 
 # nflfastR 2.2.1